Testing delete functionality.

+++ app.py
@@ -329,8 +329,6 @@ def shifter():
 @login_required
 def deleter():
     if request.method == "POST":
-        flash("Song deleted.")
-
         # Get song file path
         song = request.form.get("song")
 
@@ -339,21 +337,26 @@ def deleter():
 
         # Isolate song name
         song_name = song.rsplit(".", 1)[0]
+        
+        # Remove files if this user is the only user with this song in their library
+        if db.execute("SELECT * FROM songs WHERE song=? AND user_id=?", song, session["user_id"]) == 1:
+            # Remove uploaded song file
+            if song in [song for song in listdir(UPLOAD_FOLDER)]:
+                os.system("rm -f {}".format(UPLOAD_FOLDER + "/" + song))
 
+            # Remove stems
+            if song_name in [song for song in listdir(STEM_FOLDER)]:
+                os.system("rm -rf {}".format(STEM_FOLDER + "/" + song_name))
+
+            # Remove pitched stems
+            if song_name in [song for song in listdir(PITCHED_FOLDER)]:
+                os.system("rm -rf {}".format(PITCHED_FOLDER + "/" + song_name))
+        
+        
         # Remove song from database
-        db.execute("DELETE FROM songs WHERE song=?", song)
-
-        # Remove uploaded song file
-        if song in [song for song in listdir(UPLOAD_FOLDER)]:
-            os.system("rm -f {}".format(UPLOAD_FOLDER + "/" + song))
-
-        # Remove stems
-        if song_name in [song for song in listdir(STEM_FOLDER)]:
-            os.system("rm -rf {}".format(STEM_FOLDER + "/" + song_name))
-
-        # Remove pitched stems
-        if song_name in [song for song in listdir(PITCHED_FOLDER)]:
-            os.system("rm -rf {}".format(PITCHED_FOLDER + "/" + song_name))
+        db.execute("DELETE FROM songs WHERE song=? AND user_id=?", song, session["user_id"])
+        
+        flash("Song deleted.")
 
         return redirect("/mysongs")
